roles and permision crud

+++ app/Http/Controllers/Admin/UserController.php
@@ -11,18 +11,34 @@ class UserController extends Controller
 {
     public function index()
     {
+        // Check permission using gate
+        // $this->authorize('view_users', auth('admin')->user());
+
         $users = User::latest()->paginate(2);
+        
         return Inertia::render('admin/users/Index',[
             'users' => $users,
+            'permissions' => [
+                'can_create' => auth('admin')->user()->can('create_users'),
+                'can_edit' => auth('admin')->user()->can('edit_users'),
+                'can_delete' => auth('admin')->user()->can('delete_users'),
+                'can_view' => auth('admin')->user()->can('view_users'),
+            ]
         ]);
     }
 
     public function create()
     {
+        // Check permission using gate
+        //$this->authorize('create_users', auth('admin')->user());
+
         return Inertia::render('admin/users/Create');
     }
 
     public function store(Request $request){
+        // Check permission using gate
+        $this->authorize('create_users', auth('admin')->user());
+
         $data = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'email', 'unique:users,email'],
@@ -37,12 +53,22 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        // Check permission using gate
+        $this->authorize('edit_users', auth('admin')->user());
+
         return Inertia::render('admin/users/Edit',[
             'user' => $user,
+            'permissions' => [
+                'can_edit' => auth('admin')->user()->can('edit_users'),
+                'can_delete' => auth('admin')->user()->can('delete_users'),
+            ]
         ]);
     }
 
     public function update(Request $request, User $user){
+        // Check permission using gate
+        $this->authorize('edit_users', auth('admin')->user());
+
         $data = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'email', 'unique:users,email,'.$user->id],
@@ -62,6 +88,9 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        // Check permission using gate
+        $this->authorize('delete_users', auth('admin')->user());
+
         $user->delete();
         return redirect()->route('admin.users.index')->with('message', 'User deleted successfully');
     }
